# micro-optimization for generic tuple allocation