Not a huge savings, but I wanted to have this available for another reason too, and seemed simple enough to be worthwhile:
```julia
julia> x = 1; @btime (x, x);
  before: 69.130 ns (1 allocation: 32 bytes)
  after: 53.253 ns (1 allocation: 32 bytes)

julia> x = 1; @btime (x, x, x, x, x, x);
  before: 135.513 ns (1 allocation: 64 bytes)
  after: 115.842 ns (1 allocation: 64 bytes)
````